separate opcode_witness
