+++ stwo_cairo_prover/crates/prover/src/witness/mod.rs
@@ -1,1 +1,2 @@
 pub mod cairo;
+pub mod opcodes;
